Refactor UI into templates and API router
