+++ templates/index.html
@@ -1,0 +1,109 @@
+<html lang='en'>
+<head>
+    <meta charset='utf-8'/>
+    <title>P2P Storage UI</title>
+    <link href='https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css' rel='stylesheet'>
+</head>
+<body class='container py-4'>
+    <h1 class='mb-4'>P2P Storage Client</h1>
+
+    <section id='register' class='mb-5'>
+        <h2>Register</h2>
+        <form id='registerForm' class='row g-3'>
+            <div class='col-md-4'>
+                <label class='form-label'>Client ID</label>
+                <input type='text' class='form-control' id='regId' required>
+            </div>
+            <div class='col-md-4'>
+                <label class='form-label'>Endpoint</label>
+                <input type='text' class='form-control' id='regEndpoint' placeholder='127.0.0.1:9001' required>
+            </div>
+            <div class='col-md-4'>
+                <label class='form-label'>Available Space (MB)</label>
+                <input type='number' class='form-control' id='regSpace' required>
+            </div>
+            <div class='col-12'>
+                <button class='btn btn-primary' type='submit'>Register</button>
+            </div>
+        </form>
+    </section>
+
+    <section id='offers' class='mb-5'>
+        <h2>Offers</h2>
+        <form id='offersForm' class='row g-3 mb-3'>
+            <div class='col-md-3'>
+                <label class='form-label'>Minimum Space (MB)</label>
+                <input type='number' class='form-control' id='minSpace' value='0'>
+            </div>
+            <div class='col-md-3 align-self-end'>
+                <button class='btn btn-secondary' type='submit'>Refresh</button>
+            </div>
+        </form>
+        <ul id='offersList' class='list-group'></ul>
+    </section>
+
+    <section id='reserve' class='mb-5'>
+        <h2>Reserve Space</h2>
+        <form id='reserveForm' class='row g-3'>
+            <div class='col-md-4'>
+                <label class='form-label'>My ID</label>
+                <input type='text' class='form-control' id='reserveFrom' required>
+            </div>
+            <div class='col-md-4'>
+                <label class='form-label'>Peer</label>
+                <select id='reserveTo' class='form-select' required></select>
+            </div>
+            <div class='col-md-4'>
+                <label class='form-label'>Amount (MB)</label>
+                <input type='number' class='form-control' id='reserveAmount' required>
+            </div>
+            <div class='col-12'>
+                <button class='btn btn-primary' type='submit'>Reserve</button>
+            </div>
+        </form>
+        <div id='reserveResult' class='mt-3'></div>
+    </section>
+
+    <section id='requests' class='mb-5'>
+        <h2>Incoming Requests</h2>
+        <form id='requestsForm' class='row g-3 mb-3'>
+            <div class='col-md-3'>
+                <label class='form-label'>My ID</label>
+                <input type='text' class='form-control' id='requestsId' required>
+            </div>
+            <div class='col-md-3'>
+                <label class='form-label'>Local Port</label>
+                <input type='number' class='form-control' id='requestsPort' value='9001' required>
+            </div>
+            <div class='col-md-3 align-self-end'>
+                <button class='btn btn-secondary' type='submit'>Refresh</button>
+            </div>
+        </form>
+        <ul id='requestsList' class='list-group'></ul>
+    </section>
+
+    <section id='send' class='mb-5' style='display:none;'>
+        <h2>Send File</h2>
+        <form id='sendForm' class='row g-3'>
+            <input type='hidden' id='sendReservationId'>
+            <div class='col-md-4'>
+                <label class='form-label'>My ID</label>
+                <input type='text' class='form-control' id='sendFrom' required>
+            </div>
+            <div class='col-md-4'>
+                <label class='form-label'>Local Port</label>
+                <input type='number' class='form-control' id='sendPort' value='9001' required>
+            </div>
+            <div class='col-md-4'>
+                <label class='form-label'>File Path</label>
+                <input type='text' class='form-control' id='filePath' required>
+            </div>
+            <div class='col-12'>
+                <button class='btn btn-primary' type='submit'>Send</button>
+            </div>
+        </form>
+    </section>
+
+    <script src="/static/js/main.js"></script>
+</body>
+</html>
